# lib/api-zod/src/generated/types/registerPushTokenRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterPushTokenRequestPlatform } from "./registerPushTokenRequestPlatform";

export interface RegisterPushTokenRequest {
  deviceId: string;
  token: string;
  platform: RegisterPushTokenRequestPlatform;
}